update to Elasticsearch 6.3.1

+++ elasticsearch-test-framework/src/main/java/org/elasticsearch/test/rest/ESRestTestCase.java
@@ -2,13 +2,18 @@ package org.elasticsearch.test.rest;
 
 import org.apache.http.Header;
 import org.apache.http.HttpHost;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.client.methods.HttpHead;
+import org.apache.http.client.methods.HttpPost;
+import org.apache.http.client.methods.HttpPut;
 import org.apache.http.entity.ContentType;
 import org.apache.http.entity.StringEntity;
 import org.apache.http.message.BasicHeader;
 import org.apache.http.nio.conn.ssl.SSLIOSessionStrategy;
 import org.apache.http.ssl.SSLContexts;
-import org.apache.lucene.util.IOUtils;
+import org.elasticsearch.core.internal.io.IOUtils;
 import org.elasticsearch.action.admin.cluster.node.tasks.list.ListTasksAction;
+import org.elasticsearch.action.admin.indices.settings.put.UpdateSettingsResponse;
 import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseException;
 import org.elasticsearch.client.RestClient;
@@ -19,9 +24,12 @@ import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.common.util.concurrent.ThreadContext;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentHelper;
 import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.common.xcontent.XContentType;
 import org.elasticsearch.common.xcontent.json.JsonXContent;
+import org.elasticsearch.common.xcontent.support.XContentMapValues;
+import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.test.ESTestCase;
 import org.junit.After;
 import org.junit.AfterClass;
@@ -119,9 +127,11 @@ public abstract class ESRestTestCase extends ESTestCase {
      */
     @After
     public final void cleanUpCluster() throws Exception {
-        wipeCluster();
-        waitForClusterStateUpdatesToFinish();
-        logIfThereAreRunningTasks();
+        if (preserveClusterUponCompletion() == false) {
+            wipeCluster();
+            waitForClusterStateUpdatesToFinish();
+            logIfThereAreRunningTasks();
+        }
     }
 
     @AfterClass
@@ -147,6 +157,17 @@ public abstract class ESRestTestCase extends ESTestCase {
      */
     protected static RestClient adminClient() {
         return adminClient;
+    }
+
+    /**
+     * Returns whether to preserve the state of the cluster upon completion of this test. Defaults to false. If true, overrides the value of
+     * {@link #preserveIndicesUponCompletion()}, {@link #preserveTemplatesUponCompletion()}, {@link #preserveReposUponCompletion()}, and
+     * {@link #preserveSnapshotsUponCompletion()}.
+     *
+     * @return true if the state of the cluster should be preserved
+     */
+    protected boolean preserveClusterUponCompletion() {
+        return false;
     }
 
     /**
@@ -264,7 +285,7 @@ public abstract class ESRestTestCase extends ESTestCase {
 
         if (mustClear) {
             adminClient().performRequest("PUT", "/_cluster/settings", emptyMap(), new StringEntity(
-                clearCommand.string(), ContentType.APPLICATION_JSON));
+                Strings.toString(clearCommand), ContentType.APPLICATION_JSON));
         }
     }
 
@@ -403,7 +424,7 @@ public abstract class ESRestTestCase extends ESTestCase {
         return runningTasks;
     }
 
-    protected void assertOK(Response response) {
+    protected static void assertOK(Response response) {
         assertThat(response.getStatusLine().getStatusCode(), anyOf(equalTo(200), equalTo(201)));
     }
 
@@ -412,7 +433,7 @@ public abstract class ESRestTestCase extends ESTestCase {
      * in an non green state
      * @param index index to test for
      **/
-    protected void ensureGreen(String index) throws IOException {
+    protected static void ensureGreen(String index) throws IOException {
         Map<String, String> params = new HashMap<>();
         params.put("wait_for_status", "green");
         params.put("wait_for_no_relocating_shards", "true");
@@ -425,7 +446,7 @@ public abstract class ESRestTestCase extends ESTestCase {
      * waits until all shard initialization is completed. This is a handy alternative to ensureGreen as it relates to all shards
      * in the cluster and doesn't require to know how many nodes/replica there are.
      */
-    protected void ensureNoInitializingShards() throws IOException {
+    protected static void ensureNoInitializingShards() throws IOException {
         Map<String, String> params = new HashMap<>();
         params.put("wait_for_no_initializing_shards", "true");
         params.put("timeout", "70s");
@@ -433,22 +454,80 @@ public abstract class ESRestTestCase extends ESTestCase {
         assertOK(client().performRequest("GET", "_cluster/health/", params));
     }
 
-    protected void createIndex(String name, Settings settings) throws IOException {
+    protected static void createIndex(String name, Settings settings) throws IOException {
         createIndex(name, settings, "");
     }
 
-    protected void createIndex(String name, Settings settings, String mapping) throws IOException {
+    protected static void createIndex(String name, Settings settings, String mapping) throws IOException {
         assertOK(client().performRequest("PUT", name, Collections.emptyMap(),
             new StringEntity("{ \"settings\": " + Strings.toString(settings)
                 + ", \"mappings\" : {" + mapping + "} }", ContentType.APPLICATION_JSON)));
     }
 
-    protected void updateIndexSetting(String index, Settings.Builder settings) throws IOException {
+    protected static void updateIndexSetting(String index, Settings.Builder settings) throws IOException {
         updateIndexSetting(index, settings.build());
     }
 
-    private void updateIndexSetting(String index, Settings settings) throws IOException {
+    private static void updateIndexSetting(String index, Settings settings) throws IOException {
         assertOK(client().performRequest("PUT", index + "/_settings", Collections.emptyMap(),
             new StringEntity(Strings.toString(settings), ContentType.APPLICATION_JSON)));
+    }
+
+
+    protected static Map<String, Object> getIndexSettings(String index) throws IOException {
+        Map<String, String> params = new HashMap<>();
+        params.put("flat_settings", "true");
+        Response response = client().performRequest(HttpGet.METHOD_NAME, index + "/_settings", params);
+        assertOK(response);
+        try (InputStream is = response.getEntity().getContent()) {
+            return XContentHelper.convertToMap(XContentType.JSON.xContent(), is, true);
+        }
+    }
+
+    protected static boolean indexExists(String index) throws IOException {
+        Response response = client().performRequest(HttpHead.METHOD_NAME, index);
+        return RestStatus.OK.getStatus() == response.getStatusLine().getStatusCode();
+    }
+
+    protected static void closeIndex(String index) throws IOException {
+        Response response = client().performRequest(HttpPost.METHOD_NAME, index + "/_close");
+        assertThat(response.getStatusLine().getStatusCode(), equalTo(RestStatus.OK.getStatus()));
+    }
+
+    protected static void openIndex(String index) throws IOException {
+        Response response = client().performRequest(HttpPost.METHOD_NAME, index + "/_open");
+        assertThat(response.getStatusLine().getStatusCode(), equalTo(RestStatus.OK.getStatus()));
+    }
+
+    protected static boolean aliasExists(String alias) throws IOException {
+        Response response = client().performRequest(HttpHead.METHOD_NAME, "/_alias/" + alias);
+        return RestStatus.OK.getStatus() == response.getStatusLine().getStatusCode();
+    }
+
+    protected static boolean aliasExists(String index, String alias) throws IOException {
+        Response response = client().performRequest(HttpHead.METHOD_NAME, "/" + index + "/_alias/" + alias);
+        return RestStatus.OK.getStatus() == response.getStatusLine().getStatusCode();
+    }
+
+    @SuppressWarnings("unchecked")
+    protected static Map<String, Object> getAlias(final String index, final String alias) throws IOException {
+        String endpoint = "/_alias";
+        if (false == Strings.isEmpty(index)) {
+            endpoint = index + endpoint;
+        }
+        if (false == Strings.isEmpty(alias)) {
+            endpoint = endpoint + "/" + alias;
+        }
+        Map<String, Object> getAliasResponse = getAsMap(endpoint);
+        return (Map<String, Object>)XContentMapValues.extractValue(index + ".aliases." + alias, getAliasResponse);
+    }
+
+    protected static Map<String, Object> getAsMap(final String endpoint) throws IOException {
+        Response response = client().performRequest(HttpGet.METHOD_NAME, endpoint);
+        XContentType entityContentType = XContentType.fromMediaTypeOrFormat(response.getEntity().getContentType().getValue());
+        Map<String, Object> responseEntity = XContentHelper.convertToMap(entityContentType.xContent(),
+                response.getEntity().getContent(), false);
+        assertNotNull(responseEntity);
+        return responseEntity;
     }
 }
